+++ linuxdoom-1.10/m_misc.c
@@ -186,6 +186,7 @@ extern int	key_menu;
 extern int	mousebfire;
 extern int	mousebstrafe;
 extern int	mousebforward;
+extern int	novert;
 extern int	grabMouse;
 
 extern int	joybfire;
@@ -280,6 +281,7 @@ default_t	defaults[] =
     {"mouseb_fire",&mousebfire,0},
     {"mouseb_strafe",&mousebstrafe,1},
     {"mouseb_forward",&mousebforward,2},
+    {"novert",&novert,1},
     {"grab_mouse",&grabMouse,0},
 
     {"use_joystick",&usejoystick, 0},